migration修正

+++ backend/db/migrate/20220502055619_create_nutrients_intake_standards.rb
@@ -3,7 +3,7 @@
 class CreateNutrientsIntakeStandards < ActiveRecord::Migration[6.1]
   def change
     create_table :nutrients_intake_standards do |t|
-      t.integer :nutrients_id
+      t.bigint :nutrients_id
       t.integer :content
       t.integer :unit
       t.integer :age
