+++ backend/db/migrate/20220625073327_add_parent_id_to_to_nutrients.rb
@@ -2,6 +2,6 @@
 
 class AddParentIdToToNutrients < ActiveRecord::Migration[6.1]
   def change
-    add_column :nutrients, :parent_id, :integer
+    add_column :nutrients, :parent_id, :bigint
   end
 end
