+++ backend/db/migrate/20230226075129_create_nutrients_relations.rb
@@ -3,8 +3,8 @@
 class CreateNutrientsRelations < ActiveRecord::Migration[6.1]
   def change
     create_table :nutrients_relations do |t|
-      t.integer :parent_id
-      t.integer :child_id
+      t.bigint :parent_id
+      t.bigint :child_id
 
       t.timestamps
     end
