+++ backend/db/migrate/20230307130457_create_tag_categories.rb
@@ -3,8 +3,8 @@
 class CreateTagCategories < ActiveRecord::Migration[6.1]
   def change
     create_table :tag_categories do |t|
-      t.integer :tag_id
-      t.integer :category_id
+      t.bigint :tag_id
+      t.bigint :category_id
 
       t.timestamps
     end
